Document string schema custom attributes
This commit adds changes to document string schema custom attributes. Community at truecharts raised a concern where they are not sure how the regex works for valid_char string attribute, so the changes introduced in this commit address the concerns and while we are here document other custom string options as well.