feat(jupyter): Add Jupyter (#3616)
* Initial work for Jupyter notebook

* Fixed some things I forgot to change from base

* Added readme and changelog

* Update charts/incubator/jupyter/values.yaml

* Update charts/incubator/jupyter/values.yaml

* remove autogenerated content

* lint

* fix persistence and service

* Add some config options and try to fix permissions

I added some of the config options supported by the container. I also tried to fix permissions by setting UID and GID, but this doesn't seem to have worked

* Changed container to run as root

I changed the pod security so that it would run as root in order for it to deploy with proper folder permissions.  From my understanding while the pod is running as root the actual jupyter process is running as user 568 since I set the uid and gid to 568.

* update IDs

* some improvements

* typo

* update image repo

* update questions

* this is not needed, as users can select auto perms on persistance section

* Update charts/incubator/jupyter/questions.yaml

Signed-off-by: Stavros Kois <47820033+stavros-k@users.noreply.github.com>

* Update charts/incubator/jupyter/Chart.yaml

Signed-off-by: Stavros Kois <47820033+stavros-k@users.noreply.github.com>

Signed-off-by: Stavros Kois <47820033+stavros-k@users.noreply.github.com>
Co-authored-by: Kjeld Schouten-Lebbing <kjeld@schouten-lebbing.nl>
Co-authored-by: Stavros Kois <47820033+stavros-k@users.noreply.github.com>
Co-authored-by: Stavros kois <s.kois@outlook.com>