fix(plausible): Remove deprecated config options (#7810)
**Description**

Plausible v1.5.1 introduces changes that break the current chart. This
PR adds the required changes.

> **New first launch**
> 
> Admin user registration is now done through the UI instead of
configuration variables. At first launch, users are presented with a
registration page which will create the admin account. Admin account
related configuration variables are ignored going forward. See below for
details.
> 
> Admin user registration is required from 1.5 onwards. We are
deprecating the DISABLE_AUTH option which allowed Plausible to be used
without any user account at all.
> 
> Users with existing admin accounts are not affected, but we recommend
removing the unused configuration variables. A benefit of moving admin
account management to the UI is that you can now change the admin email
account through the UI. Previously, this was done by manual database
changes and configuration change which required restarting the service.

More info:

- https://github.com/plausible/analytics/releases/tag/v1.5.1
- https://github.com/plausible/hosting/pull/67

⚒️ Fixes  # <!--(issue)-->

**⚙️ Type of change**

- [ ] ⚙️ Feature/App addition
- [x] 🪛 Bugfix
- [ ] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 🔃 Refactor of current code

**🧪 How Has This Been Tested?**
<!--
Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce. Please also list any relevant details
for your test configuration
-->

**📃 Notes:**
<!-- Please enter any other relevant information here -->

**✔️ Checklist:**

- [x] ⚖️ My code follows the style guidelines of this project
- [x] 👀 I have performed a self-review of my own code
- [ ] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [ ] 📄 I have made corresponding changes to the documentation
- [x] ⚠️ My changes generate no new warnings
- [ ] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [x] ⬆️ I increased versions for any altered app according to semantic
versioning

**➕ App addition**

If this PR is an app addition please make sure you have done the
following.

- [ ] 🪞 I have opened a PR on
[truecharts/containers](https://github.com/truecharts/containers) adding
the container to TrueCharts mirror repo.
- [ ] 🖼️ I have added an icon in the Chart's root directory called
`icon.png`

---

_Please don't blindly check all the boxes. Read them and only check
those that apply.
Those checkboxes are there for the reviewer to see what is this all
about and
the status of this PR with a quick glance._