fix(lldap): lldap fixes and refactor (#7869)
**Description**
<!--
Please include a summary of the change and which issue is fixed. Please
also include relevant motivation and context. List any dependencies that
are required for this change.
-->
Installed lldap as it's recently come through. Addressing issues with
answers to the questions.yaml not feeding through also removed the
password var as the default can be changed in the web portal. LDAP
service also wasn't being created so addressed that too. Need to push a
seperate change to container repo to switch to a more stable image i.e.
stable instead of latest tag.

Also taken the opportunity to pare down the questions.yaml to ensure it
keeps the app as simple as possible in line with it's keep things as
simple as possible philosophy.

⚒️ Fixes  # <!--(issue)-->

**⚙️ Type of change**

- [ ] ⚙️ Feature/App addition
- [X] 🪛 Bugfix
- [ ] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [X] 🔃 Refactor of current code

**🧪 How Has This Been Tested?**
<!--
Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce. Please also list any relevant details
for your test configuration
-->

Tested initial app on my own live system and identified issues with the
questions.yaml (default in values.yaml not overridden) as well as the
ldap service not appearing.

**📃 Notes:**
<!-- Please enter any other relevant information here -->
Additional container PR to be raised changing app image tag to stable
instead of latest (potentially unstable changes with this tag that
aren't production ready).

**✔️ Checklist:**

- [X] ⚖️ My code follows the style guidelines of this project
- [X] 👀 I have performed a self-review of my own code
- [ ] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [ ] 📄 I have made corresponding changes to the documentation
- [ ] ⚠️ My changes generate no new warnings
- [ ] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [X] ⬆️ I increased versions for any altered app according to semantic
versioning

**➕ App addition**

If this PR is an app addition please make sure you have done the
following.

- [ ] 🪞 I have opened a PR on
[truecharts/containers](https://github.com/truecharts/containers) adding
the container to TrueCharts mirror repo.
- [ ] 🖼️ I have added an icon in the Chart's root directory called
`icon.png`

---

_Please don't blindly check all the boxes. Read them and only check
those that apply.
Those checkboxes are there for the reviewer to see what is this all
about and
the status of this PR with a quick glance._