chore(ci): Improved lint speed (#8809)
**Description**
<!--
Please include a summary of the change and which issue is fixed. Please
also include relevant motivation and context. List any dependencies that
are required for this change.
-->
⚒️ Fixes  # <!--(issue)-->

TODO: Switch to scripted checks (To replace ct-list's tasks):
- [x] Helm Lint
- [x] Version Checking
- [x] Run the lint inside the devcontainer (so it have the below tools
available)
- [x] YAML Schema check on Chart.yaml
- [ ] Maintener check on Chart.yaml (Moved to another PR in the future)
- [x] YAML Lint on Chart.yaml
- [x] YAML Lint on values.yaml
- [x] Sort output of lint results (failed on the bottom)
- [x] Combine verify deps in the lint job, since the verification runs
inside the fetch dep (which runs on lint job too)
- [x] Remove one of the 2 "get changed" functions (one based on tags
other on commit)
After that, move linting to single step 


https://github.com/helm/chart-testing/blob/main/pkg/tool/linter.go
https://github.com/helm/chart-testing/blob/main/pkg/tool/account.go

**⚙️ Type of change**

- [ ] ⚙️ Feature/App addition
- [ ] 🪛 Bugfix
- [ ] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] 🔃 Refactor of current code

**🧪 How Has This Been Tested?**
<!--
Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce. Please also list any relevant details
for your test configuration
-->

**📃 Notes:**
<!-- Please enter any other relevant information here -->

**✔️ Checklist:**

- [x] ⚖️ My code follows the style guidelines of this project
- [x] 👀 I have performed a self-review of my own code
- [x] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [x] 📄 I have made corresponding changes to the documentation
- [x] ⚠️ My changes generate no new warnings
- [x] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [x] ⬆️ I increased versions for any altered app according to semantic
versioning

**➕ App addition**

If this PR is an app addition please make sure you have done the
following.

- [ ] 🪞 I have opened a PR on
[truecharts/containers](https://github.com/truecharts/containers) adding
the container to TrueCharts mirror repo.
- [ ] 🖼️ I have added an icon in the Chart's root directory called
`icon.png`

---

_Please don't blindly check all the boxes. Read them and only check
those that apply.
Those checkboxes are there for the reviewer to see what is this all
about and
the status of this PR with a quick glance._